+++ src/common/logging/backend.cpp
@@ -6,7 +6,9 @@
 #include <exception>
 #include <thread>
 #include <vector>
-#include "common/common_paths.h"
+
+#include <fmt/format.h>
+
 #ifdef _WIN32
 #include <share.h>   // For _SH_DENYWR
 #include <windows.h> // For OutputDebugStringW
@@ -22,12 +24,14 @@
 #define CITRA_LINUX_GCC_BACKTRACE
 #endif
 
+#include "common/common_paths.h"
 #include "common/file_util.h"
 #include "common/literals.h"
 #include "common/settings.h"
 #include "common/thread.h"
 #include "common/logging/backend.h"
 #include "common/logging/log.h"
+#include "common/logging/log_entry.h"
 #include "common/logging/text_formatter.h"
 #include "common/string_util.h"
 #include "common/threadsafe_queue.h"
